@@ -18,7 +18,7 @@ dependencies:
   freezed_annotation: ^2.0.3
   rxdart: ^0.27.3
   flutter_riverpod: ^2.3.6
-  intl: ^0.17.0
+  intl: ^0.18.0
   shared_preferences: ^2.0.20
   dio: ^5.1.0
   file_picker: ^5.2.6
@@ -27,27 +27,24 @@ dependencies:
   audioplayers: ^3.0.1
   share_plus: ^6.3.4
   path_provider: ^2.0.9
-  firebase_core: ^1.13.1
-  firebase_messaging: ^11.2.6
+  firebase_core: ^2.4.0
+  firebase_messaging: ^14.0.1
   flutter_local_notifications: ^13.0.0
   url_launcher: ^6.1.10
-  ffmpeg_kit_flutter: 4.5.1-LTS
+  ffmpeg_kit_flutter: 5.1.0-LTS
   skeletons: ^0.0.3
-  video_trimmer: ^1.1.3
-  firebase_storage: ^10.2.16
-  firebase_auth: ^3.11.2
-  cloud_firestore: ^3.1.16
+  video_trimmer: ^2.1.0
+  firebase_storage: ^11.0.1
+  firebase_auth: ^4.0.0
+  cloud_firestore: ^4.0.1
   chewie: ^1.4.0
   video_player: ^2.6.1
   cross_file: ^0.3.3+2
   video_thumbnail: ^0.5.3
   twitter_login: ^4.3.1
-  app_review: ^2.1.2+1
-  package_info_plus: ^1.4.3+1
+  package_info_plus: ^3.0.0
   font_awesome_flutter: ^10.4.0
-
-dependency_overrides:
-  firebase_core_platform_interface: 4.5.1
+  in_app_review: ^2.0.6
 
 dev_dependencies:
   flutter_test:
