@@ -42,7 +42,7 @@ class TrimSoundForDetectionViewModel
   }
 
   Future<void> onPlay() async {
-    final playbackState = await state.trimmer.videPlaybackControl(
+    final playbackState = await state.trimmer.videoPlaybackControl(
       startValue: state.startValue,
       endValue: state.endValue,
     );
